refactor: 경주할 자동차 이름 입력 시 dto 전달하도록 리팩터링
